refactor(pivot): move pivot model in util/types lib
use a relative alias to point the correct model instead of the whole lib
