+++ apps/metadata-converter/src/app/components/record-form/record-form.component.ts
@@ -3,7 +3,6 @@ import {
   Component,
   EventEmitter,
   Input,
-  OnInit,
   Output,
 } from '@angular/core'
 import {
@@ -17,7 +16,7 @@ import {
   ServiceOnlineResource,
   ServiceOnlineResourceType,
   ServiceRecord,
-} from '@geonetwork-ui/metadata-converter'
+} from '@geonetwork-ui/util/types/metadata'
 
 @Component({
   selector: 'gn-ui-record-form',
