+++ libs/api/metadata-converter/src/index.ts
@@ -1,2 +1,1 @@
 export * from './lib/iso19139/converter'
-export * from './lib/model'
