+++ libs/api/metadata-converter/src/lib/iso19139/converter.ts
@@ -1,4 +1,8 @@
-import { CatalogRecord, DatasetRecord, ServiceRecord } from '../model'
+import {
+  CatalogRecord,
+  DatasetRecord,
+  ServiceRecord,
+} from '@geonetwork-ui/util/types/metadata'
 import {
   createDocument,
   createElement,
