+++ libs/util/types/src/lib/metadata/index.ts
@@ -1,0 +1,3 @@
+export * from './contact.model'
+export * from './organisation.model'
+export * from './metadata.model'
