+++ libs/util/types/src/lib/metadata/organisation.model.ts
@@ -1,0 +1,6 @@
+export interface Organization {
+  name: string
+  description?: string
+  website?: URL
+  logoUrl?: URL
+}
